+++ gazebo/sensors/CameraSensor.cc
@@ -25,6 +25,7 @@
 #include "gazebo/common/Events.hh"
 #include "gazebo/common/Exception.hh"
 #include "gazebo/common/Image.hh"
+#include "gazebo/common/CommonIface.hh"
 
 #include "gazebo/msgs/msgs.hh"
 
@@ -75,7 +76,17 @@ std::string CameraSensor::Topic() const
 {
   std::string topicName = "~/";
   topicName += this->ParentName() + "/" + this->Name() + "/image";
-  boost::replace_all(topicName, "::", "/");
+  common::replaceAll(topicName, "::", "/");
+
+  return topicName;
+}
+
+//////////////////////////////////////////////////
+std::string CameraSensor::TopicIgn() const
+{
+  std::string topicName = this->ScopedName() + "/image";
+  common::replaceAll(topicName, "::", "/");
+  common::replaceAll(topicName, " ", "_");
 
   return topicName;
 }
@@ -84,8 +95,12 @@ std::string CameraSensor::Topic() const
 void CameraSensor::Load(const std::string &_worldName)
 {
   Sensor::Load(_worldName);
-  this->imagePub = this->node->Advertise<msgs::ImageStamped>(
-      this->Topic(), 50);
+  this->imagePub = this->node->Advertise<msgs::ImageStamped>(this->Topic(), 50);
+
+  ignition::transport::AdvertiseMessageOptions opts;
+  opts.SetMsgsPerSec(50);
+  this->imagePubIgn = this->nodeIgn.Advertise<ignition::msgs::ImageStamped>(
+      this->TopicIgn(), opts);
 }
 
 //////////////////////////////////////////////////
@@ -200,22 +215,48 @@ bool CameraSensor::UpdateImpl(const bool /*_force*/)
 
   this->camera->PostRender();
 
-  if (this->imagePub && this->imagePub->HasConnections())
+
+  if ((this->imagePub && this->imagePub->HasConnections()) ||
+      this->imagePubIgn.HasConnections())
   {
-    msgs::ImageStamped msg;
-    msgs::Set(msg.mutable_time(), this->scene->SimTime());
-    msg.mutable_image()->set_width(this->camera->ImageWidth());
-    msg.mutable_image()->set_height(this->camera->ImageHeight());
-    msg.mutable_image()->set_pixel_format(common::Image::ConvertPixelFormat(
-          this->camera->ImageFormat()));
+    auto simTime = this->scene->SimTime();
+    if (this->imagePub && this->imagePub->HasConnections())
+    {
+      msgs::ImageStamped msg;
+      msgs::Set(msg.mutable_time(), simTime);
+      msg.mutable_image()->set_width(this->camera->ImageWidth());
+      msg.mutable_image()->set_height(this->camera->ImageHeight());
+      msg.mutable_image()->set_pixel_format(common::Image::ConvertPixelFormat(
+            this->camera->ImageFormat()));
 
-    msg.mutable_image()->set_step(this->camera->ImageWidth() *
-        this->camera->ImageDepth());
-    msg.mutable_image()->set_data(this->camera->ImageData(),
-        msg.image().width() * this->camera->ImageDepth() *
-        msg.image().height());
+      msg.mutable_image()->set_step(this->camera->ImageWidth() *
+          this->camera->ImageDepth());
+      msg.mutable_image()->set_data(this->camera->ImageData(),
+          msg.image().width() * this->camera->ImageDepth() *
+          msg.image().height());
 
-    this->imagePub->Publish(msg);
+      this->imagePub->Publish(msg);
+    }
+
+    if (this->imagePubIgn.HasConnections())
+    {
+      ignition::msgs::ImageStamped msg;
+      msg.mutable_time()->set_sec(simTime.sec);
+      msg.mutable_time()->set_nsec(simTime.nsec);
+
+      msg.mutable_image()->set_width(this->camera->ImageWidth());
+      msg.mutable_image()->set_height(this->camera->ImageHeight());
+      msg.mutable_image()->set_pixel_format(common::Image::ConvertPixelFormat(
+            this->camera->ImageFormat()));
+
+      msg.mutable_image()->set_step(this->camera->ImageWidth() *
+          this->camera->ImageDepth());
+      msg.mutable_image()->set_data(this->camera->ImageData(),
+          msg.image().width() * this->camera->ImageDepth() *
+          msg.image().height());
+
+      this->imagePubIgn.Publish(msg);
+    }
   }
 
   this->dataPtr->rendered = false;
@@ -280,7 +321,8 @@ bool CameraSensor::SaveFrame(const std::string &_filename)
 bool CameraSensor::IsActive() const
 {
   return Sensor::IsActive() ||
-    (this->imagePub && this->imagePub->HasConnections());
+    (this->imagePub && this->imagePub->HasConnections()) ||
+    this->imagePubIgn.HasConnections();
 }
 
 //////////////////////////////////////////////////